llmservice/requirements.txt
flask
torch
torchaudio
ffmpeg-python
python-dotenv
torchtune==0.4.0
torchao==0.9.0
transformers==4.49.0
tokenizers==0.21.0
huggingface-hub==0.28.1
silentcipher @ git+https://github.com/SesameAILabs/silentcipher@master
moshi==0.2.2
gunicorn>=21.2.0
guidance